@@ -1,13 +1,14 @@
-import {Response} from '@angular/http'
 import {Observable} from 'rxjs/Observable'
 import 'rxjs/add/observable/throw'
+import { HttpResponse, HttpErrorResponse } from '@angular/common/http';
 
 export class ErrorHandler {
-  static handleError(error: Response | any){
+  static handleError(error: HttpErrorResponse | any) {
     let errorMessage: string
-    if (error instanceof Response){
-      errorMessage = `${error.url}: ${error.status} - ${error.statusText || ''}`;
-    }else{
+    if (error instanceof HttpErrorResponse) {
+      const body = error.error;
+      errorMessage = `${error.url}: ${error.status} - ${error.statusText || ''} ${body}`;
+    }else {
       errorMessage = error.message ? error.message : error.toString();
     }
     console.log(errorMessage)
