@@ -5,7 +5,8 @@ export class Order {
         public number: number,
         public optionalAddress: string,
         public paymentOption: string,
-        public orderItems: OrderItem[] = []
+        public orderItems: OrderItem[] = [],
+        public id?: string
 ) {}
 }
 
