@@ -3,14 +3,14 @@ import { Injectable } from '@angular/core';
 import { CartItem } from '../restaurant-detail/shopping-cart/cart-item.model';
 import { Order } from './order.model';
 import { Observable } from 'rxjs/Observable';
-import { Http, RequestOptions, Headers } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
 import { MEAT_API } from '../app.api';
 
 @Injectable()
 export class OrderService {
 
     constructor(private cartService: ShoppingCartService,
-                private http: Http) {
+                private http: HttpClient) {
 
     }
 
@@ -34,16 +34,12 @@ export class OrderService {
         return this.cartService.total();
     }
 
-    clear(){
+    clear() {
         this.cartService.clear();
     }
 
     checkOrder(order: Order): Observable<string> {
-        const headers: Headers = new Headers();
-        headers.append('Content-Type', 'application/json');
-
-        return this.http.post(`${MEAT_API}/orders`, JSON.stringify(order), new RequestOptions({headers: headers}))
-            .map(response => response.json())
+        return this.http.post<Order>(`${MEAT_API}/orders`, order)
             .map(order => order.id);
     }
 }
