@@ -1,8 +1,8 @@
-export interface MenuItem{
+export interface MenuItem {
     id: string;
     imagePath: string;
     name: string;
     description: string;
     price: number;
     restaurantId: string;
-}
+};
