@@ -1,39 +1,34 @@
 import { Restaurant } from './restaurant/restaurant.model';
 import { Injectable } from '@angular/core';
-import { Http } from '@angular/http';
 import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/catch';
 import { MEAT_API } from '../app.api';
-import { ErrorHandler } from '../app.error-handler';
 import { MenuItem } from '../restaurant-detail/menu-item/menu-item.model';
+import { HttpClient, HttpParams } from '@angular/common/http';
 
 @Injectable()
 export class RestaurantsService {
 
-    constructor(private http: Http) { }
+    constructor(private http: HttpClient) { }
 
-    restaurants(): Observable<Restaurant[]> {
-       return this.http.get(`${MEAT_API}/restaurants`)
-            .map(response => response.json())
-            .catch(ErrorHandler.handleError);
+    restaurants(search?: string): Observable<Restaurant[]> {
+        let params: HttpParams = undefined;
+        if (search) {
+            params = new HttpParams().set('q', search);
+        }
+       return this.http.get<Restaurant[]>(`${MEAT_API}/restaurants`, {params: params});
     }
 
     restaurantById(id: string): Observable<Restaurant> {
-        return this.http.get(`${MEAT_API}/restaurants/${id}`)
-            .map(response => response.json())
-            .catch(ErrorHandler.handleError);
+        return this.http.get<Restaurant>(`${MEAT_API}/restaurants/${id}`);
     }
 
     menuOfRestaurant(id: string): Observable<MenuItem[]> {
-        return this.http.get(`${MEAT_API}/restaurants/${id}/menu`)
-          .map(response => response.json())
-          .catch(ErrorHandler.handleError);
+        return this.http.get<MenuItem[]>(`${MEAT_API}/restaurants/${id}/menu`);
       }
 
       reviewsOfRestaurant(id: string): Observable<any> {
-        return this.http.get(`${MEAT_API}/restaurants/${id}/reviews`)
-          .map(response => response.json())
-          .catch(ErrorHandler.handleError);
+        return this.http.get(`${MEAT_API}/restaurants/${id}/reviews`);
       }
 }
